pyproject.toml
[project]
name = "datasette-demo-dbs"
version = "0.1"
description = "Fetch missing demo DBs from URLs when Datasette first starts"
readme = "README.md"
authors = [{name = "Simon Willison"}]
license = "Apache-2.0"
classifiers=[
    "Framework :: Datasette"
]
requires-python = ">=3.9"
dependencies = [
    "datasette"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/datasette/datasette-demo-dbs"
Changelog = "https://github.com/datasette/datasette-demo-dbs/releases"
Issues = "https://github.com/datasette/datasette-demo-dbs/issues"
CI = "https://github.com/datasette/datasette-demo-dbs/actions"

[project.entry-points.datasette]
demo_dbs = "datasette_demo_dbs"

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio", "pytest-httpx"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
